REFACTOR(handle_delete): refactor handle_get by dividing handling based on has_trailing_slash Condition

+++ src/methods/method_delete.cpp
@@ -10,13 +10,21 @@
 
 void	Connection::is_dir_Case(void)
 {
+	if (!std::filesystem::exists(_full_path)) {
+		P_DEBUG("2nd 404 in is_dir_Case");
+		// set_full_status_code(409);
+		set_full_status_code(404);
+		return ;
+	}
 	if (!std::filesystem::is_empty(_full_path)) {
 		// Verzeichnis ist nicht leer, daher kann es nicht gelöscht werden – Conflict
 		set_full_status_code(409);
-	} else {
+	}
+	else {
 		if (rmdir(_full_path.c_str()) == 0) {
 			set_full_status_code(204);
-		} else {
+		}
+		else {
 			//TODO: add more precise error checks with errno
 			set_full_status_code(403);
 		}
@@ -25,14 +33,37 @@ void	Connection::is_dir_Case(void)
 
 void	Connection::is_file_Case(void)
 {
-	if (unlink(_full_path.c_str()) == 0) {
-		// Löschung erfolgreich: 204 No Content
-		set_full_status_code(204);
-	} else {
-		//TODO: add more precise error checks with errno
-		// Löschen schlug fehl (z.B. Berechtigungsproblem): 403 Forbidden
-		set_full_status_code(403);
+	if (!std::filesystem::exists(_full_path)) {
+		P_DEBUG("3nd 404 in is_file_Case");
+		set_full_status_code(404);
+		return ;
 	}
+	if (std::filesystem::is_regular_file(_full_path))
+	{
+		if (unlink(_full_path.c_str()) == 0) {
+			// Löschung erfolgreich: 204 No Content
+			set_full_status_code(204);
+		}
+		else {
+			//TODO: add more precise error checks with errno
+			// Löschen schlug fehl (z.B. Berechtigungsproblem): 403 Forbidden
+			set_full_status_code(403);
+		}
+	}
+	else if (std::filesystem::is_directory(_full_path))
+	{
+		P_DEBUG("conflict case");
+		set_full_status_code(409);
+		return ;
+	}
+	// if (unlink(_full_path.c_str()) == 0) {
+	// 	// Löschung erfolgreich: 204 No Content
+	// 	set_full_status_code(204);
+	// } else {
+	// 	//TODO: add more precise error checks with errno
+	// 	// Löschen schlug fehl (z.B. Berechtigungsproblem): 403 Forbidden
+	// 	set_full_status_code(403);
+	// }
 }
 
 /*
@@ -62,20 +93,30 @@ void	Connection::handle_delete(void)
 	P_DEBUG("alo");
 	try
 	{
-		if (!std::filesystem::exists(_full_path)) {
-			std::cout << "In doesnt exist  Case\n";
-			set_full_status_code(404); // Ressource existiert nicht
-			return;
-		}
-		else if (std::filesystem::is_regular_file(_full_path)) {
-			std::cout << "In is_regular_file() Case\n";
-			is_file_Case();
-		}
-		else if (std::filesystem::is_directory(_full_path)) {
-			std::cout << "In is_directory() Case\n";
+		bool has_trailing_slash = \
+			!request.uri->path.empty() && request.uri->path.back() == '/';
+		// if (!std::filesystem::exists(_full_path)) {
+		// 	std::cout << "In doesnt exist  Case\n";
+		// 	P_DEBUG("1er 404");
+		// 	set_full_status_code(404); // Ressource existiert nicht
+		// 	return;
+		// }
+		if(has_trailing_slash)
+		{
 			is_dir_Case();
 		}
-		// Falls die Ressource weder Datei noch Verzeichnis ist, gib 404 zurück:
+		else if (!has_trailing_slash) {
+			is_file_Case();
+		}
+		// else if (std::filesystem::is_regular_file(_full_path)) {
+		// 	std::cout << "In is_regular_file() Case\n";
+		// 	is_file_Case();
+		// }
+		// else if (std::filesystem::is_directory(_full_path)) {
+		// 	std::cout << "In is_directory() Case\n";
+		// 	is_dir_Case();
+		// }
+		// // Falls die Ressource weder Datei noch Verzeichnis ist, gib 404 zurück:
 		else {
 			set_full_status_code(404);
 		}
